Make screenshot generation work again
This took 3 days to figure out.
It involved creating an Ubuntu 20.04 virtualbox,
compiling and modifying multiple times the latest Mesa,
packaging and uploading it as a tar.gz to openrndr/setup-opengl,
tweaks and debugging suggested by Edwin until discovering that
skipping the openSWR renderer altogether seems
to solve the issue.