[orx-processing] Fix PShape conversion when multiple ShapeContours are present
Add a demo to show it works.

I tested this in Processing and noticed
it's necessary to avoid wrapping the first vertices in `s.beginContour()` / `s.endContour()`, otherwise
it does not render properly there.

I also noticed that `createShape()` in Processing creates a `PShape(PShape.GEOMETRY)` by default, but we create a `PShape.PATH`. I tried changing this but it failed to work without further tweaks, so I kept it as `PATH` for now.